final commit

+++ Database/Procedures.sql
@@ -1,4 +1,3 @@
-
 --Admin
 ---------
 --Add Admin procedure
@@ -24,6 +23,8 @@ as
 begin
 if exists(select Email from AdminTable where Email = @email)
 select Email, Password from AdminTable where Email = @email and Password = @password
+else
+Print 'Invalid'
 end;
 go
 
@@ -37,6 +38,7 @@ end;
 go
 
 ------------------------------------------------------------------------------
+
 --User
 -----------
 --Add user
@@ -51,14 +53,15 @@ AS
 BEGIN
     IF EXISTS(SELECT Email FROM UserTable WHERE Email = @email)
     BEGIN
+        PRINT 'Invalid'
     END
     ELSE
     BEGIN
         INSERT INTO UserTable (Email, Password, UserName, MobileNumber, UserRole) 
-        VALUES (@email, @password, @username, @mobileNumber, @userRole);
-    END;
+        VALUES (@email, @password, @username, @mobileNumber, @userRole)
+    END
 END;
-
+go
 
 --User Login 
 --------------
@@ -67,6 +70,8 @@ as
 begin
 if exists(select Email from UserTable where  Email =@email)
 select Email, Password from UserTable where Email = @email and Password = @password
+else
+Print 'Invalid'
 end;
 go
 
@@ -84,7 +89,7 @@ go
 create procedure getAllUsers
 as
 begin
-select UserId, Email, Password, UserName, MobileNumber, UserRole from UserTable
+select * from UserTable
 end;
 go
 
@@ -120,8 +125,7 @@ go
 create procedure deleteUsers
 as
 begin
-delete from UserTable 
-where UserId > 0
+delete from UserTable where UserId>0
 end;
 go
 
@@ -161,8 +165,7 @@ go
 create proc getAllServiceCenterDetails
 as
 begin
-select serviceCenterId, serviceCenterName, serviceCenterPhone, serviceCenterAddress, serviceCenterImageUrl, serviceCenterMailId,
-serviceCost, serviceCenterStartTime, serviceCenterEndTime, serviceCenterDescription from AddCenters
+select * from AddCenters
 end ;
 go
 
@@ -303,8 +306,7 @@ go
 create procedure getAllAppointments
 as
 begin
-select ID, customerName, productName, productModelNo, dateofPurchase, contactNumber, problemDescription, bookedSlots,
-dateOfAppointment, email, serviceCenterId, serviceCenterName, dateOfAppointmentBooking, serviceCost from Appointments
+select * from Appointments
 end;
 go
 ------------------------------------------------------------------------------
@@ -429,6 +431,6 @@ go
 create proc GetAllReviews
 as
 begin
-select userEmail, userName, serviceCenterId, Rating, review from ServiceReviews
+select * from ServiceReviews
 end;
 go
